requirements.txt
autogen-agentchat
autogen-core
autogen-ext
asyncio
tiktoken
autogen-ext[docker]
pydantic
typing
opeai
python-dotenv
langchain
langchain-core
langchain-groq
langchain-chroma
langchain-community
langchain-huggingface
langchain-text-splitters
sentence_transformers
ipykernel
pypdf
mysql-connector-python